update: added server, reorganized files, fixed webpack, rendering text from server to browser
