+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
 .env
-dist
+client/dist/bundle.js
