+++ client/src/index.jsx
@@ -1,5 +1,5 @@
 import ReactDOM from "react-dom/client";
-import App from "./App.jsx";
+import App from "./components/App.jsx";
 
 const root = ReactDOM.createRoot(
   document.getElementById("app")
