+++ package.json
@@ -4,7 +4,8 @@
   "description": "adoption application",
   "main": "index.js",
   "scripts": {
-    "build": "webpack --watch"
+    "build": "webpack --watch",
+    "start": "nodemon server/index.js"
   },
   "repository": {
     "type": "git",
