+++ server/db/index.js
@@ -1,0 +1,1 @@
+/// optional database server file
